+++ src/Firebase/Auth/CreateActionLink/GuzzleApiClientHandler.php
@@ -6,22 +6,27 @@ namespace Kreait\Firebase\Auth\CreateActionLink;
 
 use GuzzleHttp\ClientInterface;
 use GuzzleHttp\Exception\GuzzleException;
+use GuzzleHttp\Psr7\Request;
+use GuzzleHttp\Psr7\Utils;
 use InvalidArgumentException;
 use Kreait\Firebase\Auth\CreateActionLink;
 use Kreait\Firebase\Util\JSON;
+use Psr\Http\Message\RequestInterface;
 
 final class GuzzleApiClientHandler implements Handler
 {
     private ClientInterface $client;
+    private string $projectId;
 
-    public function __construct(ClientInterface $client)
+    public function __construct(ClientInterface $client, string $projectId)
     {
         $this->client = $client;
+        $this->projectId = $projectId;
     }
 
     public function handle(CreateActionLink $action): string
     {
-        $request = new ApiRequest($action);
+        $request = $this->createRequest($action);
 
         try {
             $response = $this->client->send($request, ['http_errors' => false]);
@@ -44,5 +49,30 @@ final class GuzzleApiClientHandler implements Handler
         }
 
         return (string) $actionCode;
+    }
+
+    private function createRequest(CreateActionLink $action): RequestInterface
+    {
+        $data = \array_filter([
+            'requestType' => $action->type(),
+            'email' => $action->email(),
+            'returnOobLink' => true,
+        ]) + $action->settings()->toArray();
+
+        if ($tenantId = $action->tenantId()) {
+            $uri = "https://identitytoolkit.googleapis.com/v1/projects/{$this->projectId}/tenants/{$tenantId}/accounts:sendOobCode";
+        } else {
+            $uri = "https://identitytoolkit.googleapis.com/v1/projects/{$this->projectId}/accounts:sendOobCode";
+        }
+
+        $body = Utils::streamFor(JSON::encode($data, JSON_FORCE_OBJECT));
+
+        $headers = \array_filter([
+            'Content-Type' => 'application/json; charset=UTF-8',
+            'Content-Length' => (string) $body->getSize(),
+            'X-Firebase-Locale' => $action->locale(),
+        ]);
+
+        return new Request('POST', $uri, $headers, $body);
     }
 }