+++ src/Firebase/Auth/CreateSessionCookie/GuzzleApiClientHandler.php
@@ -6,21 +6,26 @@ namespace Kreait\Firebase\Auth\CreateSessionCookie;
 
 use GuzzleHttp\ClientInterface;
 use GuzzleHttp\Exception\GuzzleException;
+use GuzzleHttp\Psr7\Request;
+use GuzzleHttp\Psr7\Utils;
 use Kreait\Firebase\Auth\CreateSessionCookie;
 use Kreait\Firebase\Util\JSON;
+use Psr\Http\Message\RequestInterface;
 
 final class GuzzleApiClientHandler implements Handler
 {
     private ClientInterface $client;
+    private string $projectId;
 
-    public function __construct(ClientInterface $client)
+    public function __construct(ClientInterface $client, string $projectId)
     {
         $this->client = $client;
+        $this->projectId = $projectId;
     }
 
     public function handle(CreateSessionCookie $action): string
     {
-        $request = new ApiRequest($action);
+        $request = $this->createRequest($action);
 
         try {
             $response = $this->client->send($request, ['http_errors' => false]);
@@ -46,5 +51,28 @@ final class GuzzleApiClientHandler implements Handler
         }
 
         throw new FailedToCreateSessionCookie($action, $response, 'The response did not contain a session cookie');
+    }
+
+    private function createRequest(CreateSessionCookie $action): RequestInterface
+    {
+        $data = [
+            'idToken' => $action->idToken(),
+            'validDuration' => $action->ttlInSeconds(),
+        ];
+
+        if ($tenantId = $action->tenantId()) {
+            $uri = "https://identitytoolkit.googleapis.com/v1/projects/{$this->projectId}/tenants/{$tenantId}:createSessionCookie";
+        } else {
+            $uri = "https://identitytoolkit.googleapis.com/v1/projects/{$this->projectId}:createSessionCookie";
+        }
+
+        $body = Utils::streamFor(JSON::encode($data, JSON_FORCE_OBJECT));
+
+        $headers = \array_filter([
+            'Content-Type' => 'application/json; charset=UTF-8',
+            'Content-Length' => (string) $body->getSize(),
+        ]);
+
+        return new Request('POST', $uri, $headers, $body);
     }
 }