+++ tests/Integration/Database/AuthVariableOverrideTest.php
@@ -32,6 +32,7 @@ final class AuthVariableOverrideTest extends DatabaseTestCase
     public function testItCanAccessAReferenceThatBelongsToTheSameUser(): void
     {
         $uid = $this->auth->signInAnonymously()->firebaseUserId();
+        \assert(\is_string($uid));
 
         $this->publishRules(__FUNCTION__, ['.read' => 'auth.uid === "'.$uid.'"']);
 
@@ -41,13 +42,14 @@ final class AuthVariableOverrideTest extends DatabaseTestCase
             $db->getReference(self::$refPrefix)->getChild(__FUNCTION__)->getValue();
             $this->addToAssertionCount(1);
         } finally {
-            $this->deleteUser($uid);
+            $this->auth->deleteUser($uid);
         }
     }
 
     public function testItCanNotAccessAReferenceThatRequiresAnotherUser(): void
     {
         $uid = $this->auth->signInAnonymously()->firebaseUserId();
+        \assert(\is_string($uid));
 
         $this->publishRules(__FUNCTION__, ['.read' => 'auth.uid === "someone-else"']);
 
@@ -57,30 +59,34 @@ final class AuthVariableOverrideTest extends DatabaseTestCase
             $this->expectException(PermissionDenied::class);
             $db->getReference(self::$refPrefix)->getChild(__FUNCTION__)->getValue();
         } finally {
-            $this->deleteUser($uid);
+            $this->auth->deleteUser($uid);
         }
     }
 
     public function testItCanAccessAPublicReferenceWhenAuthOverrideIsSetToBeUnauthenticated(): void
     {
         $uid = $this->auth->signInAnonymously()->firebaseUserId();
+        \assert(\is_string($uid));
 
         $this->publishRules(__FUNCTION__, ['.read' => true]);
 
         try {
-            $db = $this->databaseWithAuthOverride(null);
+            $this->databaseWithAuthOverride(null)
+                ->getReference(self::$refPrefix)
+                ->getChild(__FUNCTION__)
+                ->getValue()
+            ;
 
-            $ref = $db->getReference(self::$refPrefix)->getChild(__FUNCTION__);
-            $ref->getValue();
             $this->addToAssertionCount(1);
         } finally {
-            $this->deleteUser($uid);
+            $this->auth->deleteUser($uid);
         }
     }
 
     public function testWhenUnauthenticatedItCanNotAccessAReferenceThatRequiresAuthentication(): void
     {
         $uid = $this->auth->signInAnonymously()->firebaseUserId();
+        \assert(\is_string($uid));
 
         $this->publishRules(__FUNCTION__, ['.read' => 'auth != null']);
 
@@ -90,7 +96,7 @@ final class AuthVariableOverrideTest extends DatabaseTestCase
             $this->expectException(PermissionDenied::class);
             $db->getReference(self::$refPrefix)->getChild(__FUNCTION__)->getValue();
         } finally {
-            $this->deleteUser($uid);
+            $this->auth->deleteUser($uid);
         }
     }