+++ tests/Integration/Request/UpdateUserTest.php
@@ -178,12 +178,14 @@ final class UpdateUserTest extends IntegrationTestCase
     {
         $user = $this->auth->createAnonymousUser();
 
-        $this->assertNull($user->metadata->passwordUpdatedAt);
+        try {
+            $this->assertNull($user->metadata->passwordUpdatedAt);
 
-        $updatedUser = $this->auth->updateUser($user->uid, ['password' => 'new-password']);
+            $updatedUser = $this->auth->updateUser($user->uid, ['password' => 'new-password']);
 
-        $this->assertInstanceOf(\DateTimeImmutable::class, $updatedUser->metadata->passwordUpdatedAt);
-
-        $this->deleteUser($user);
+            $this->assertInstanceOf(\DateTimeImmutable::class, $updatedUser->metadata->passwordUpdatedAt);
+        } finally {
+            $this->auth->deleteUser($user->uid);
+        }
     }
 }