+++ tests/Integration/TenantAwareAuthTest.php
@@ -4,19 +4,24 @@ declare(strict_types=1);
 
 namespace Kreait\Firebase\Tests\Integration;
 
-use Kreait\Firebase\Contract\Auth;
-use Kreait\Firebase\Tests\IntegrationTestCase;
+use Kreait\Firebase\Auth\CreateSessionCookie\FailedToCreateSessionCookie;
 
 /**
  * @internal
  */
-final class TenantAwareAuthTest extends IntegrationTestCase
+final class TenantAwareAuthTest extends AuthTestCase
 {
-    private Auth $auth;
-
     protected function setUp(): void
     {
         $this->auth = self::$factory->withTenantId(self::TENANT_ID)->createAuth();
+    }
+
+    public function testCreateSessionCookie(): void
+    {
+        $this->expectException(FailedToCreateSessionCookie::class);
+        $this->expectExceptionMessageMatches('/UNSUPPORTED_TENANT_OPERATION/');
+
+        parent::testCreateSessionCookie();
     }
 
     public function testNewUsersAreScopedToATenant(): void